+++ Codigo/services/problemas.py
@@ -1,74 +1,87 @@
 PROBLEMAS_MOCK = [
     {
-        "id": 1,
+        "id": "prob-001",
         "titulo": "Suma de dos números",
         "descripcion": "Dado dos números enteros, retorna su suma.",
-        "casos_de_prueba": [
+        "casos_prueba": [
             {
-                "id": 1,
+                "id": "caso-001",
                 "descripcion": "Suma básica positiva",
-                "input": "3 5",
-                "expected_output": "8"
+                "entrada": "3 5",
+                "salida_esperada": "8",
+                "es_publico": True,
+                "orden": 1
             },
             {
-                "id": 2,
+                "id": "caso-002",
                 "descripcion": "Suma con número negativo",
-                "input": "-2 4",
-                "expected_output": "2"
+                "entrada": "-2 4",
+                "salida_esperada": "2",
+                "es_publico": True,
+                "orden": 2
             }
         ]
     },
     {
-        "id": 2,
+        "id": "prob-002",
         "titulo": "Número mayor",
         "descripcion": "Dado dos números enteros, retorna el mayor.",
-        "casos_de_prueba": [
+        "casos_prueba": [
             {
-                "id": 1,
+                "id": "caso-003",
                 "descripcion": "Primer número mayor",
-                "input": "9 3",
-                "expected_output": "9"
+                "entrada": "9 3",
+                "salida_esperada": "9",
+                "es_publico": True,
+                "orden": 1
             },
             {
-                "id": 2,
+                "id": "caso-004",
                 "descripcion": "Segundo número mayor",
-                "input": "4 7",
-                "expected_output": "7"
+                "entrada": "4 7",
+                "salida_esperada": "7",
+                "es_publico": True,
+                "orden": 2
             }
         ]
     },
-    
     {
-        "id": 3,
+        "id": "prob-003",
         "titulo": "Palíndromo",
-        "descripcion": "Dada una cadena de texto, retorna True si es un palíndromo (se lee igual de izquierda a derecha que de derecha a izquierda) y False en caso contrario.",
-        "casos_de_prueba": [
+        "descripcion": "Dada una cadena de texto, retorna True si es un palíndromo y False en caso contrario.",
+        "casos_prueba": [
             {
-                "id": 1,
+                "id": "caso-005",
                 "descripcion": "Palíndromo simple",
-                "input": "radar",
-                "expected_output": "True"
+                "entrada": "radar",
+                "salida_esperada": "True",
+                "es_publico": True,
+                "orden": 1
             },
             {
-                "id": 2,
+                "id": "caso-006",
                 "descripcion": "No es palíndromo",
-                "input": "python",
-                "expected_output": "False"
+                "entrada": "python",
+                "salida_esperada": "False",
+                "es_publico": True,
+                "orden": 2
             },
             {
-                "id": 3,
+                "id": "caso-007",
                 "descripcion": "Palíndromo con mayúsculas",
-                "input": "Ana",
-                "expected_output": "True"
+                "entrada": "Ana",
+                "salida_esperada": "True",
+                "es_publico": True,
+                "orden": 3
             }
         ]
     }
 ]
 
 
-def get_problema_por_id(problem_id):
+def get_problema_por_id(problema_id):
     for problema in PROBLEMAS_MOCK:
-        if problema["id"] == problem_id:
+        if problema["id"] == problema_id:
             return problema
     return None
